requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
beanie==1.23.6
motor==3.3.2
pymongo==4.6.0

# Validation & Security
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Utils
python-multipart==0.0.6
pyyaml==6.0.1

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
httpx==0.25.2
